requirements.txt
Flask
llama_index
python-dotenv
httpx
transformers
